code-snippets/python/rag-pipeline/requirements.txt
# QueryLex RAG Pipeline - Dependencies
# Install with: pip install -r requirements.txt

# === Core (Required) ===
tiktoken>=0.5.0                # Token counting for chunk sizing
openai>=1.0.0                  # LLM client (also used for Mistral, DeepSeek via base_url)
numpy>=1.24.0                  # Vector operations
requests>=2.28.0               # HTTP client for embedding APIs
python-dotenv>=1.0.0           # Load .env file for API keys

# === Vector Store (Supabase + pgvector) ===
supabase>=2.0.0                # Supabase Python client
httpx>=0.24.0                  # HTTP client with timeout support (used by Supabase)

# === Embeddings ===
sentence-transformers>=2.2.0   # Cross-encoder reranking model
torch>=2.0.0                   # PyTorch (required by sentence-transformers)

# === Document Processing ===
langchain-community>=0.0.10    # HuggingFace embeddings wrapper
tqdm>=4.65.0                   # Progress bars for batch processing

# === Optional: PDF Processing ===
# Needed only if using document_processor.process_pdf()
# unstructured[pdf]>=0.10.0    # Advanced PDF parsing with element detection
# PyPDF2>=3.0.0                # Fallback PDF parser
# python-docx>=0.8.11          # Word document processing

# === Optional: Additional Embedding Providers ===
# cohere>=4.0.0                # Cohere embed-v3
# voyageai>=0.1.0              # Voyage AI embeddings

# === Optional: NLTK (for legal tokenizer) ===
# nltk>=3.8.0                  # Natural language toolkit
